Category filters - Fix category filter - New locales
